+++ hutool-log/src/test/java/cn/hutool/log/test/StaticLogTest.java
@@ -1,12 +1,10 @@
 package cn.hutool.log.test;
 
 import cn.hutool.log.LogFactory;
-import cn.hutool.log.dialect.console.ConsoleColorLog;
+import cn.hutool.log.StaticLog;
 import cn.hutool.log.dialect.console.ConsoleColorLogFactory;
 import cn.hutool.log.dialect.console.ConsoleLogFactory;
 import org.junit.Test;
-
-import cn.hutool.log.StaticLog;
 
 public class StaticLogTest {
 	@Test
@@ -24,7 +22,5 @@ public class StaticLogTest {
 		StaticLog.error("This is static {} log", "error");
 		StaticLog.warn("This is static {} log", "warn");
 		StaticLog.trace("This is static {} log", "trace");
-		ConsoleColorLog.setWarnColor(31);
-		StaticLog.warn("This is static {} log", "warn");
 	}
 }